+++ tests/screenshot/apps/components/ToggleButton.js
@@ -3,6 +3,13 @@ import ToggleButton from '../../../../ToggleButton';
 const ToggleButtonTests = [
 	<ToggleButton>Click me!</ToggleButton>,
 	<ToggleButton underline>Click me!</ToggleButton>,
+	<ToggleButton underline defaultSelected>Click me!</ToggleButton>,
+	<ToggleButton underline size="huge">Click me!</ToggleButton>,
+	<ToggleButton underline size="huge" defaultSelected>Click me!</ToggleButton>,
+	<ToggleButton underline size="small">Click me!</ToggleButton>,
+	<ToggleButton underline size="small" defaultSelected>Click me!</ToggleButton>,
+	<ToggleButton underline size="smallest">Click me!</ToggleButton>,
+	<ToggleButton underline size="smallest" defaultSelected>Click me!</ToggleButton>,
 	<ToggleButton disabled>Click me!</ToggleButton>,
 
 	// size